phishing-email-analyzer/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
openai==1.28.0
google-generativeai==0.3.0
python-dotenv==1.0.0
pydantic==2.6.1
pydantic-settings==2.1.0
